fixed add to wishlist/add to cart, added routes for getting/editing users and orders

+++ Controllers/ordersController.js
@@ -13,6 +13,17 @@ const getUserOrders = async (req, res) => {
   }
 };
 
+// GET ALL ORDERS
+const getAllOrders = async (req, res) => {
+  try {
+    const orders = await Order.find();
+
+    res.json({ success: true, orders });
+  } catch (err) {
+    res.json({ message: err });
+  }
+};
+
 // CREATE ORDER
 const createOrder = async (req, res) => {
   try {
@@ -21,8 +32,6 @@ const createOrder = async (req, res) => {
     if (!cart) {
       return res.status(404).json({ message: "Cart not found" });
     }
-
-    const existingOrder = await Order.find({ user: req.user._id });
 
     let totalPrice = 0;
 
@@ -38,16 +47,31 @@ const createOrder = async (req, res) => {
           name: item.product,
           modelNumber: item.modelNumber,
           quantity: item.quantity,
+          regularPrice: item.regularPrice,
+          salePrice: item.salePrice,
           totalPrice: item.price,
         },
       };
     };
 
+    let orderSavings = 0;
+
+    cart.map((item) => {
+      if (item.regularPrice > item.salePrice) {
+        orderSavings += (item.regularPrice - item.salePrice) * item.quantity;
+      }
+    });
+
+    console.log(orderSavings);
+
     const newOrder = new Order({
       user: req.user._id,
+      firstName: req.user.firstName,
+      lastName: req.user.lastName,
       orderStatus: "pending",
       orderItems: [...cart.map(orderItems)],
       orderTotal: totalPrice,
+      orderSavings,
     });
 
     await Order.create(newOrder);
@@ -57,6 +81,26 @@ const createOrder = async (req, res) => {
     res.json({ message: "Order created", newOrder });
   } catch (error) {
     res.status(400).json({ message: error.message });
+  }
+};
+
+// UPDATE ORDER STATUS
+const updateOrderStatus = async (req, res) => {
+  const existingOrder = await Order.findOne({ orderNumber: req.params.id });
+
+  if (!existingOrder) {
+    return res.status(404).json({ message: "Order not found" });
+  }
+
+  if (req.user.scope !== "admin") {
+    return res.status(401).json({ message: "Not authorized" });
+  } else {
+    const updatedOrder = await Order.updateOne(existingOrder, {
+      ...Order,
+      orderStatus: req.body.orderStatus,
+    });
+
+    res.json({ message: "Order updated", updatedOrder });
   }
 };
 
@@ -86,7 +130,13 @@ const deleteUserOrder = async (req, res) => {
 // DELETE ALL USER ORDERS
 const deleteAllUserOrders = async (req, res) => {
   try {
-    const orders = await Order.deleteMany({ user: req.user._id });
+    let orders;
+
+    if (req.user.scope !== "admin") {
+      orders = await Order.deleteMany({ user: req.user._id });
+    } else {
+      return res.status(401).json({ message: "Not authorized" });
+    }
 
     res.json({ message: "All orders deleted", orders });
   } catch (err) {
@@ -96,7 +146,9 @@ const deleteAllUserOrders = async (req, res) => {
 
 module.exports = {
   getUserOrders,
+  getAllOrders,
   createOrder,
+  updateOrderStatus,
   deleteUserOrder,
   deleteAllUserOrders,
 };
